update: pkg/apperror
